payment-python-fastapi/payment-service/requirements.txt
CouchDB==1.2
fastapi==0.54.1
kafka-python==2.0.1